Telemetry: implement collection of execution/transfer times for resources

+++ edgeless_node/src/wasmi_runner/guest_api_binding.rs
@@ -32,8 +32,8 @@ pub fn cast_raw(
     payload_len: i32,
 ) -> Result<(), wasmi::core::Trap> {
     let mem = get_memory(&mut caller)?;
-    let node_id = mem.data_mut(&mut caller)[instance_node_id_ptr as usize..(instance_node_id_ptr as usize) + 16 as usize].to_vec();
-    let component_id = mem.data_mut(&mut caller)[instance_component_id_ptr as usize..(instance_component_id_ptr as usize) + 16 as usize].to_vec();
+    let node_id = mem.data_mut(&mut caller)[instance_node_id_ptr as usize..(instance_node_id_ptr as usize) + 16_usize].to_vec();
+    let component_id = mem.data_mut(&mut caller)[instance_component_id_ptr as usize..(instance_component_id_ptr as usize) + 16_usize].to_vec();
     let instance_id = edgeless_api::function_instance::InstanceId {
         node_id: uuid::Uuid::from_bytes(node_id.try_into().map_err(|_| wasmi::core::Trap::new("uuid error"))?),
         function_id: uuid::Uuid::from_bytes(component_id.try_into().map_err(|_| wasmi::core::Trap::new("uuid error"))?),
@@ -57,8 +57,8 @@ pub fn call_raw(
 ) -> Result<i32, wasmi::core::Trap> {
     let mem = get_memory(&mut caller)?;
     let alloc = get_alloc(&mut caller)?;
-    let node_id = mem.data_mut(&mut caller)[instance_node_id_ptr as usize..(instance_node_id_ptr as usize) + 16 as usize].to_vec();
-    let component_id = mem.data_mut(&mut caller)[instance_component_id_ptr as usize..(instance_component_id_ptr as usize) + 16 as usize].to_vec();
+    let node_id = mem.data_mut(&mut caller)[instance_node_id_ptr as usize..(instance_node_id_ptr as usize) + 16_usize].to_vec();
+    let component_id = mem.data_mut(&mut caller)[instance_component_id_ptr as usize..(instance_component_id_ptr as usize) + 16_usize].to_vec();
     let instance_id = edgeless_api::function_instance::InstanceId {
         node_id: uuid::Uuid::from_bytes(node_id.try_into().map_err(|_| wasmi::core::Trap::new("uuid error"))?),
         function_id: uuid::Uuid::from_bytes(component_id.try_into().map_err(|_| wasmi::core::Trap::new("uuid error"))?),
